--- lib/widgets/user_transactions.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-import './new_transaction.dart';
-import './transaction_list.dart';
-import '../models/transaction.dart';
-
-class UserTransactions extends StatefulWidget {
-  @override
-  _UserTransactionsState createState() => _UserTransactionsState();
-}
-
-class _UserTransactionsState extends State<UserTransactions> {
-  final List<Transaction> _userTransactions = [
-    new Transaction(
-      id: 't1',
-      title: 'New shoes',
-      amount: 69.99,
-      date: DateTime.now(),
-    ),
-    new Transaction(
-      id: 't2',
-      title: 'Groceries',
-      amount: 16.53,
-      date: DateTime.now(),
-    ),
-  ];
-
-  void _addNewTransaction(String txTitle, double txAmount) {
-    final newTx = Transaction(
-      title: txTitle,
-      amount: txAmount,
-      date: DateTime.now(),
-      id: DateTime.now().toString(),
-    );
-
-    setState(() {
-      _userTransactions.add(newTx);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: <Widget>[
-        NewTransaction(_addNewTransaction),
-        TransactionList(_userTransactions),
-      ],
-    );
-  }
-}
